update1

+++ database/migrations/2022_04_29_213119_create_products_table.php
@@ -13,19 +13,23 @@ class CreateProductsTable extends Migration
      */
     public function up()
     {
-        Schema::create('products', function (Blueprint $table) {
-            $table->id();
-            $table->string("title");
-            $table->string("slug",191)->unique();
-            $table->string("description");
-            $table->decimal("price",8,2)->default(0);
-            $table->decimal("old_price",8,2)->default(0);
-            $table->decimal("inStock")->default(0);
-            $table->string("image");
-            $table->bigInteger("category_id")->unsigned();
-            $table->timestamps();
-            $table->foreign("category_id")->references("id")->on("categories")->onDelete("cascade");
-        });
+        if(!Schema::hasTable('products')){
+            Schema::create('products', function (Blueprint $table) {
+                $table->id();
+                $table->string("title");
+               // $table->string("slug",191)->unique();
+                $table->string("description");
+                $table->decimal("price",8,2)->default(0);
+                $table->decimal("old_price",8,2)->default(0);
+                $table->decimal("inStock")->default(0);
+                $table->string("image");
+                $table->bigInteger("category_id")->unsigned();
+                $table->timestamps();
+                $table->foreign("category_id")->references("id")->on("categories")->onDelete("cascade");
+            });
+
+        }
+
     }
 
     /**
