+++ database/migrations/2022_04_29_221034_create_orders_table.php
@@ -13,21 +13,25 @@ class CreateOrdersTable extends Migration
      */
     public function up()
     {
-        Schema::create('orders', function (Blueprint $table) {
-            $table->id();
-            $table->bigInteger("user_id")->unsigned();
-            $table->string("product_name");
-            $table->integer("qty");
-            $table->decimal("price",8,2);
-            $table->total("total",8,2);
-            $table->boolean("paid")->default(0);
-            $table->boolean("delivered")->default(0);
-            $table->timestamps();
-            $table->foreign("user_id")
-                  ->references("id")
-                  ->on("users")
-                  ->onDelete("cascade");
-        });
+        if(!Schema::hasTable('orders')){
+            Schema::create('orders', function (Blueprint $table) {
+                $table->id();
+                $table->bigInteger("user_id")->unsigned();
+                $table->string("product_name");
+                $table->integer("qty");
+                $table->decimal("price",8,2);
+                $table->decimal("total",8,2);
+                $table->boolean("paid")->default(0);
+                $table->boolean("delivered")->default(0);
+                $table->timestamps();
+                $table->foreign("user_id")
+                      ->references("id")
+                      ->on("users")
+                      ->onDelete("cascade");
+            });
+
+        }
+       
     }
 
     /**
